+++ server/src/api/jobs/instances/CompetitionStarting.ts
@@ -1,5 +1,5 @@
+import * as competitionService from '@services/internal/competitions';
 import { onCompetitionStarting } from '../../events';
-import * as competitionService from '../../modules/competitions/competition.service';
 import { Job } from '../index';
 
 class CompetitionStarting implements Job {
