+++ server/src/api/jobs/instances/RemoveFromGroupCompetitions.ts
@@ -1,4 +1,4 @@
-import { removeFromGroupCompetitions } from '../../modules/competitions/competition.service';
+import { removeFromGroupCompetitions } from '@services/internal/competitions';
 import { Job } from '../index';
 
 class RemoveFromGroupCompetitions implements Job {
