+++ server/src/api/modules/records/record.route.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import * as controller from './record.controller';
+import * as controller from '@controllers/records';
 
 const api = Router();
