+++ server/tsconfig.json
@@ -10,6 +10,11 @@
     "outDir": "dist",
     "importHelpers": true,
     "baseUrl": ".",
+    "paths": {
+      "@controllers/*": ["./src/api/controllers/*"],
+      "@services/*": ["./src/api/services/*"],
+      "@routes/*": ["./src/api/routes/*"]
+    },
     "lib": ["es6", "es2017", "esnext"]
   },
   "exclude": ["node_modules", "dist", "__tests__"]
